pyproject.toml
[project]
name = "comfyui_a1rspace"
description = "A comfyui custom node package used by myself."
version = "1.0.0"
license = {file = "LICENSE"}
dependencies = ["sefetensors", "transformers", "torch", "numpy", "einops", "Pillow", "kornia", "requests", "openai", "psutil", "huggingface-hub", "onnxruntime"]

[project.urls]
Repository = "https://github.com/a1rchan/ComfyUI_A1rSpace"

[tool.comfy]
PublisherId = "a1rchan"
DisplayName = "ComfyUI_A1rSpace"
Icon = ""
